int -> double

+++ assignment-geometric-shapes-start/app/src/geometric/Geometric.java
@@ -6,7 +6,7 @@ interface Geometric{
     public double get_bottom_rect();
     public double get_top_rect();
     public double get_area();
-    public void   move_object(int dx, int dy);
+    public void   move_object(double dx, double dy);
 }
 
 interface GeometricPredicate{
